backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
requests
pandas
numpy
scikit-learn
joblib
pyyaml
